refactor: rename "comparaisons" folder to "comparisons" to use english naming as in backend
