+++ app.territoiresentransitions.react/app/(authed)/collectivite/[collectiviteId]/(acces-restreint)/referentiel/[referentielId]/(overview)/@tabs/evolutions/page.tsx
@@ -1,4 +1,4 @@
-import { ScoreEvolutions } from '@/app/referentiels/comparaisons';
+import { ScoreEvolutions } from '@/app/referentiels/comparisons';
 
 export default function Page() {
   return <ScoreEvolutions />;
