+++ app.territoiresentransitions.react/src/app/pages/collectivite/Referentiels/DownloadScore/download-score.modal.tsx
@@ -1,5 +1,5 @@
 import { useExportComparisonScores } from '@/app/referentiels/audits/AuditComparaison/useExportComparisonScore';
-import { DownloadSnapshotsDropdown } from '@/app/referentiels/comparaisons/dropdowns/download-snapshots.dropdown';
+import { DownloadSnapshotsDropdown } from '@/app/referentiels/comparisons/dropdowns/download-snapshots.dropdown';
 import { useListSnapshots } from '@/app/referentiels/use-snapshot';
 import SpinnerLoader from '@/app/ui/shared/SpinnerLoader';
 import { ReferentielId } from '@/domain/referentiels';
