+++ app.territoiresentransitions.react/src/referentiels/comparisons/deleteSnapshot/delete-snapshot.modal.tsx
