+++ app.territoiresentransitions.react/src/referentiels/comparisons/dropdowns/evolutions-snapshots.dropdown.tsx
@@ -1,7 +1,7 @@
 import { useCurrentCollectivite } from '@/api/collectivites';
 import { RouterOutput } from '@/api/utils/trpc/client';
-import { DeleteSnapshotButton } from '@/app/referentiels/comparaisons/deleteSnapshot/delete-snapshot.button';
-import { UpdateSnapshotNameButton } from '@/app/referentiels/comparaisons/updateSnapshotName/update-snapshot-name.button';
+import { DeleteSnapshotButton } from '@/app/referentiels/comparisons/deleteSnapshot/delete-snapshot.button';
+import { UpdateSnapshotNameButton } from '@/app/referentiels/comparisons/updateSnapshotName/update-snapshot-name.button';
 import { SnapshotJalonEnum } from '@/domain/referentiels/snapshots';
 import { Icon, SelectMultiple, SelectMultipleProps } from '@/ui';
