+++ app.territoiresentransitions.react/src/referentiels/comparisons/updateSnapshotName/update-snapshot-name.button.tsx
