+++ app.territoiresentransitions.react/src/referentiels/comparisons/utils.ts
