+++ app.territoiresentransitions.react/src/referentiels/tableau-de-bord/graphs/tdb-score-total.chart.tsx
@@ -1,5 +1,5 @@
 import { ReferentielId } from '@/domain/referentiels';
-import { ScoreTotalEvolutionsChart } from '../../comparaisons/evolutions-score-total.chart';
+import { ScoreTotalEvolutionsChart } from '../../comparisons/evolutions-score-total.chart';
 import { useListSnapshots } from '../../use-snapshot';
 import { GraphCard } from './EtatDesLieuxGraphs';
